Tidy some signed/unsigned comparisons

+++ jim-signal.c
@@ -16,9 +16,9 @@
 
 #define MAX_SIGNALS_WIDE (sizeof(jim_wide) * 8)
 #if defined(NSIG)
-    #define MAX_SIGNALS ((NSIG < MAX_SIGNALS_WIDE) ? NSIG : MAX_SIGNALS_WIDE)
+    #define MAX_SIGNALS (int)((NSIG < MAX_SIGNALS_WIDE) ? NSIG : MAX_SIGNALS_WIDE)
 #else
-    #define MAX_SIGNALS MAX_SIGNALS_WIDE
+    #define MAX_SIGNALS (int)MAX_SIGNALS_WIDE
 #endif
 
 static jim_wide *sigloc;
